Back/refactoring/protocols approaches

+++ backend/src/main/kotlin/com/jetbrains/life_science/category/service/CategoryInfo.kt
@@ -7,5 +7,5 @@ interface CategoryInfo {
 
     val parentId: Long?
 
-    val order: Int
+    val order: Long
 }
