+++ backend/src/main/kotlin/com/jetbrains/life_science/category/view/CategorySubcategoryView.kt
@@ -3,5 +3,5 @@ package com.jetbrains.life_science.category.view
 data class CategorySubcategoryView(
     val id: Long?,
     val name: String,
-    val order: Int
+    val order: Long
 )
